+++ src/onegov.gazette/HISTORY.rst
@@ -10,7 +10,7 @@ Changelog
 - Changes layout of print only and at cost options.
   [msom]
 
-1.11.6 (2018-02-07)
+1.12.6 (2018-02-07)
 ~~~~~~~~~~~~~~~~~~~
 
 - Adds author fields (name, date, place) to notices.